% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_module.R
\name{plot.qshap_rsq}
\alias{plot.qshap_rsq}
\title{Plot method for qshap_rsq objects}
\usage{
\method{plot}{qshap_rsq}(
  x,
  y = NULL,
  type = c("rsq", "elbow", "cumu", "gcorr", "hist", "density", "loss", "heatmap",
    "loss_heatmap"),
  ...
)
}
\arguments{
\item{x}{A `qshap_rsq` object.}

\item{y}{Not used.}

\item{type}{Plot type: one of "rsq", "elbow", "cumu", "gcorr", "hist",
"density", "loss", or "heatmap". `"loss_heatmap"` is retained as an
alias for `"heatmap"`.}

\item{...}{Passed to the underlying visualization function.}
}
\value{
A ggplot2 object (invisibly).
}
\description{
This S3 method enables `plot(x, ...)` where `x` is a `qshap_rsq` object.
It dispatches to the visualization functions in `vis`.
}
