% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_module.R
\docType{data}
\name{vis}
\alias{vis}
\title{Visualization Module for Q-SHAP Results}
\format{
An environment with visualization functions:
\describe{
  \item{rsq}{Bar plot of feature-specific R-squared values}
  \item{elbow}{Elbow plot showing top contributing features}
  \item{cumu}{Cumulative explained variance plot}
  \item{gcorr}{Generalized correlation plot (square root of R-squared)}
  \item{hist}{Histogram of feature-specific R-squared contributions}
  \item{density}{Density plot of feature-specific R-squared contributions}
  \item{loss}{Interactive loss explorer (requires shiny)}
  \item{loss_heatmap}{Heatmap of observation-level contributions}
}
}
\usage{
vis
}
\description{
An environment containing visualization functions for Q-SHAP results.
Access functions using \code{vis$rsq()}, \code{vis$elbow()}, etc.
}
\keyword{internal}
